+++ backend/src/main/java/com/pios/repository/GarminActivityLapRepository.java
@@ -1,5 +1,6 @@
 package com.pios.repository;
 
+import com.pios.domain.Activity;
 import com.pios.domain.GarminActivityLap;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,4 +10,6 @@ import java.util.List;
 @Repository
 public interface GarminActivityLapRepository extends JpaRepository<GarminActivityLap, Long> {
     List<GarminActivityLap> findByActivityIdIn(List<Long> activityIds);
+
+    void deleteByActivity(Activity activity);
 }